correctly ocr and overlay
