+++ screen/ocr.py
@@ -1,20 +1,14 @@
 import pytesseract
-from PIL import Image
-
-
-def extract_text(image):
-    """ Extract text from an image using OCR. """
-    text = pytesseract.image_to_string(image)
-    return text
+from pytesseract import Output
 
 
 def extract_text_and_boxes(image):
     """ Extract text and bounding boxes from an image using OCR. """
-    data = pytesseract.image_to_data(
-        image, output_type=pytesseract.Output.DICT)
-
-    text_list = data['text']
-    boxes = [(data['left'][i], data['top'][i], data['width'][i], data['height'][i])
-             for i in range(len(data['text'])) if data['text'][i].strip() != '']
-
-    return text_list, boxes
+    data = pytesseract.image_to_data(image, output_type=Output.DICT)
+    text_box_pairs = []
+    for i in range(len(data['text'])):
+        if data['text'][i].strip() != '':
+            box = (data['left'][i], data['top'][i],
+                   data['width'][i], data['height'][i])
+            text_box_pairs.append((data['text'][i], box))
+    return text_box_pairs
